Major overhaul
Changed the Shapes class to include all potential matrices of every
shape and orientation. Also changed what is a shape to include the
type, orientation, matrix, and the global x y positions. I do believe
this is inefficient, but I am just doing what is working. Fixed the
rotation and block creation