Changing to enums
Using these data structures to help organize since only using a matrix
is getting confusing and messy. This cleans things up by quite a bit.
Still need to make a ‘isValid’ function to check for overlaps (have the
start of this with isPlacedHuh for a falling block, but not for when it
is being rotated or moved into a block). I should really start on
graphics soon